pyproject.toml
[project]
name = "flask-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask (==2.3.3)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg[binary] (==3.1.18)",
    "flask-sqlalchemy (==3.1.1)",
    "pytz (>=2025.2,<2026.0)",
    "flask-migrate (>=4.1.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false